Output naming schema redux
