+++ eeauditor/processor/outputs/cam_mongodb_output.py
