+++ eeauditor/processor/outputs/dynamodb_output.py
