+++ eeauditor/processor/outputs/mongodb_output.py
