+++ eeauditor/processor/outputs/sechub_output.py
