+++ eeauditor/processor/outputs/stdout_output.py
